Module pallet_asset_index::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_default_parts_6 as tt_default_parts;

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageAssets

_GeneratedPrefixForStorageDeposits

_GeneratedPrefixForStorageIndexTokenDepositRange

_GeneratedPrefixForStorageIndexTokenLocks

_GeneratedPrefixForStorageLockedIndexToken

_GeneratedPrefixForStorageLockupPeriod

_GeneratedPrefixForStoragePendingWithdrawals

_GeneratedPrefixForStorageRedemptionFee

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

WeightInfo

Trait for the asset-index pallet extrinsic weights.

Type Definitions

Assets

(AssetId) -> AssetAvailability

Deposits

All timestamped deposits of an account.

IndexTokenDepositRange

The range of the index token equivalent a deposit must be in in order to be allowed.

IndexTokenLocks

Tracks the locks of the minted index token that are locked up until their LockupPeriod is over (AccountId) -> Vec

LockedIndexToken

Tracks the amount of the currently locked index token per user. This is equal to the sum(IndexTokenLocks[AccountId]) (AccountId) -> Balance

LockupPeriod

Store a duration (in blocks) of the lockup period

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

PendingWithdrawals

All currently pending withdrawal sets for an account. Where a single PendingRedemption is the result of a withdraw call.

RedemptionFee

stores a range of redemption fee