Trait pallet_asset_index::pallet::Config

pub trait Config: Config + MaybeAssetIdConvert<u8, Self::AssetId> {

    type AdminOrigin: EnsureOrigin<Self::Origin, Success = <Self as Config>::AccountId>;
    type IndexToken: LockableCurrency<Self::AccountId, Balance = Self::Balance>;
    type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128> + BaseFee;
    type LockupPeriod: Get<Self::BlockNumber>;
    type LockupPeriodRange: LockupPeriodRange<Self::BlockNumber>;
    type IndexTokenLockIdentifier: Get<LockIdentifier>;
    type MinimumRedemption: Get<Self::Balance>;
    type WithdrawalPeriod: Get<Self::BlockNumber>;
    type RemoteAssetManager: RemoteAssetManager<Self::AccountId, Self::AssetId, Self::Balance>;
    type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize;
    type MaxActiveDeposits: Get<u32>;
    type MaxDecimals: Get<u8>;
    type RedemptionFee: Get<RedemptionFeeRange<Self::BlockNumber>>;
    type SelfAssetId: Get<Self::AssetId>;
    type Currency: MultiReservableCurrency<Self::AccountId, CurrencyId = Self::AssetId, Balance = Self::Balance>;
    type PriceFeed: PriceFeed<Self::AssetId>;
    type PriceFeedBenchmarks: PriceFeedBenchmarks<Self::AccountId, Self::AssetId>;
    type SaftRegistry: SaftRegistry<Self::AssetId, Self::Balance>;
    type BaseWithdrawalFee: Get<FeeRate>;
    type TreasuryPalletId: Get<PalletId>;
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type StringLimit: Get<u32>;
    type WeightInfo: WeightInfo;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type AdminOrigin: EnsureOrigin<Self::Origin, Success = <Self as Config>::AccountId>

Origin that is allowed to administer the index

type IndexToken: LockableCurrency<Self::AccountId, Balance = Self::Balance>

Currency implementation to use as the index token

type Balance: Parameter + Member + AtLeast32BitUnsigned + Default + Copy + MaybeSerializeDeserialize + Into<u128> + BaseFee

The balance type used within this pallet

type LockupPeriod: Get<Self::BlockNumber>

Period after the minting of the index token for which 100% is locked up. Only applies to users contributing assets directly to index

type LockupPeriodRange: LockupPeriodRange<Self::BlockNumber>

Range of the lockup period

type IndexTokenLockIdentifier: Get<LockIdentifier>

The identifier for the index token lock. Used to lock up deposits for T::LockupPeriod.

type MinimumRedemption: Get<Self::Balance>

The minimum amount of the index token that can be redeemed for the underlying asset in the index

type WithdrawalPeriod: Get<Self::BlockNumber>

Minimum amount of time between redeeming index tokens and being able to withdraw the awarded assets

type RemoteAssetManager: RemoteAssetManager<Self::AccountId, Self::AssetId, Self::Balance>

Type that handles cross chain transfers

type AssetId: Parameter + Member + Copy + MaybeSerializeDeserialize

Type used to identify assets

type MaxActiveDeposits: Get<u32>

Restricts how many deposits can be active

type MaxDecimals: Get<u8>

Restricts the max limit of decimals in metadata

type RedemptionFee: Get<RedemptionFeeRange<Self::BlockNumber>>

Determines the redemption fee in complete_withdraw

type SelfAssetId: Get<Self::AssetId>

The native asset id

type Currency: MultiReservableCurrency<Self::AccountId, CurrencyId = Self::AssetId, Balance = Self::Balance>

Currency type for deposit/withdraw assets to/from the user’s sovereign account

type PriceFeed: PriceFeed<Self::AssetId>

The types that provides the necessary asset price pairs

type PriceFeedBenchmarks: PriceFeedBenchmarks<Self::AccountId, Self::AssetId>

The type that provides benchmark features of pallet_price_feed

type SaftRegistry: SaftRegistry<Self::AssetId, Self::Balance>

The type registry that stores all NAV for non liquid assets

type BaseWithdrawalFee: Get<FeeRate>

The basic fees that apply when a withdrawal is executed

type TreasuryPalletId: Get<PalletId>

The treasury’s pallet id, used for deriving its sovereign account ID.

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type StringLimit: Get<u32>

The maximum length of a name or symbol stored on-chain.

type WeightInfo: WeightInfo

The weight for this pallet’s extrinsics.